Refactor optional Triton imports
