+++ test/test_triton.py
@@ -1,8 +1,11 @@
+import pytest
 import torch
 from torch import Tensor
 
-from pyg_lib._triton import tl, triton
-from pyg_lib.testing import onlyCUDA, onlyTriton
+from pyg_lib.testing import onlyCUDA
+
+triton = pytest.importorskip('triton')
+tl = pytest.importorskip('triton.language')
 
 
 @triton.jit
@@ -32,7 +35,6 @@ def add(x: Tensor, y: Tensor) -> Tensor:
 
 
 @onlyCUDA
-@onlyTriton
 def test_triton():
     x = torch.rand(100, device='cuda')
     y = torch.rand(100, device='cuda')
